requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-cloud-aiplatform>=1.43.0
google-cloud-vectorsearch>=0.9.0
vertexai>=1.43.0
google-generativeai>=0.3.0
pydantic>=2.5.0
gunicorn>=21.2.0
Pillow>=10.0.0
slowapi>=0.1.9
python-dotenv>=1.0.0